requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchainhub
pinecone-client
python-dotenv
pypdf
streamlit
tiktoken
requests
numpy
pandas
